refactor: OnBoardingRepository에서 온보딩 완료 여부 통신하도록 수정
