+++ android/app/src/main/java/com/created/team201/data/datasource/remote/login/AuthDataSource.kt
@@ -1,7 +1,6 @@
 package com.created.team201.data.datasource.remote.login
 
 import com.created.team201.data.remote.response.AuthResponseDto
-import com.created.team201.data.remote.response.IsOnBoardingDoneResponseDto
 
 interface AuthDataSource {
 
@@ -10,6 +9,4 @@ interface AuthDataSource {
     suspend fun getLoginValidity(token: String)
 
     suspend fun getRenewedAccessToken(token: String): AuthResponseDto
-
-    suspend fun getIsOnboardingDone(): IsOnBoardingDoneResponseDto
 }
