+++ android/app/src/main/java/com/created/team201/data/datasource/remote/login/AuthDataSourceImpl.kt
@@ -2,7 +2,6 @@ package com.created.team201.data.datasource.remote.login
 
 import com.created.team201.data.remote.api.AuthService
 import com.created.team201.data.remote.response.AuthResponseDto
-import com.created.team201.data.remote.response.IsOnBoardingDoneResponseDto
 
 class AuthDataSourceImpl(
     private val authService: AuthService,
@@ -17,10 +16,5 @@ class AuthDataSourceImpl(
 
     override suspend fun getRenewedAccessToken(token: String): AuthResponseDto {
         return authService.postRenewedAccessToken(token)
-    }
-
-
-    override suspend fun getIsOnboardingDone(): IsOnBoardingDoneResponseDto {
-        return authService.getIsOnboardingDone()
     }
 }
