+++ android/app/src/main/java/com/created/team201/presentation/login/LoginViewModel.kt
@@ -9,11 +9,14 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import com.created.domain.repository.AuthRepository
+import com.created.domain.repository.OnBoardingRepository
 import com.created.team201.application.Team201App
 import com.created.team201.data.datasource.local.TokenDataSourceImpl
+import com.created.team201.data.datasource.remote.OnBoardingDataSourceImpl
 import com.created.team201.data.datasource.remote.login.AuthDataSourceImpl
 import com.created.team201.data.remote.NetworkServiceModule
 import com.created.team201.data.repository.AuthRepositoryImpl
+import com.created.team201.data.repository.OnBoardingRepositoryImpl
 import com.created.team201.presentation.login.LoginViewModel.State.FAIL
 import com.created.team201.presentation.login.LoginViewModel.State.SUCCESS
 import com.created.team201.presentation.onBoarding.model.OnBoardingDoneState
@@ -21,6 +24,7 @@ import kotlinx.coroutines.launch
 
 class LoginViewModel(
     private val authRepository: AuthRepository,
+    private val onBoardingRepository: OnBoardingRepository,
 ) : ViewModel() {
     private val _signUpState: MutableLiveData<State> = MutableLiveData()
     val signUpState: LiveData<State> get() = _signUpState
@@ -42,7 +46,7 @@ class LoginViewModel(
 
     fun getIsOnboardingDone() {
         viewModelScope.launch {
-            authRepository.getIsOnboardingDone()
+            onBoardingRepository.getIsOnboardingDone()
                 .onSuccess { state ->
                     _onBoardingDoneState.value = OnBoardingDoneState.Success(state)
                 }.onFailure {
@@ -67,6 +71,11 @@ class LoginViewModel(
                         ),
                         TokenDataSourceImpl(Team201App.provideTokenStorage()),
                     ),
+                    OnBoardingRepositoryImpl(
+                        OnBoardingDataSourceImpl(
+                            NetworkServiceModule.onBoardingService,
+                        ),
+                    )
                 )
             }
         }
