+++ android/domain/src/main/java/com/created/domain/repository/AuthRepository.kt
@@ -9,6 +9,4 @@ interface AuthRepository {
     suspend fun requestSignIn(): Result<Unit>
 
     suspend fun renewAccessToken()
-
-    suspend fun getIsOnboardingDone(): Result<Boolean>
 }
